+++ packages/react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @lottiefiles/dotlottie-react
 
+## 0.19.5
+### Patch Changes
+
+- Updated dependencies [6252c99]
+- Updated dependencies [e322ffb]
+  - @lottiefiles/dotlottie-web@0.75.0
+
 ## 0.19.4
 ### Patch Changes